Review replication of corporate,corporate admin,corporate user, RM data to portal closes #116, closes #117, closes #119, closes #122

+++ src/main/java/org/trishanku/oa/admin/service/CustomerAdminServiceImpl.java
@@ -136,7 +136,7 @@ public class CustomerAdminServiceImpl implements CustomerAdminService{
 
         existingUserDetails.setAuthorizationDetails(jwtUtil.extractUsernameFromRequest());
         User savedUser = userRepository.save(existingUserDetails);
-        userRepository.delete(savedUser);
+        if(savedUser.isDeleteFlag()) userRepository.delete(savedUser);
         return userMapper.userToUserDTO(savedUser);
     }
